requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart
yt-dlp==2023.11.16
openai-whisper==20231117
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
torchaudio
soundfile
edge-tts 